review: rename discover/Tutorials to discover/tutorial/TutorialList to avoid confusion
