+++ apps/crn-frontend/src/discover/Routes.tsx
@@ -7,8 +7,14 @@ import { DiscoverPage } from '@asap-hub/react-components';
 const loadGuides = () =>
   import(/* webpackChunkName: "discover-guides" */ './Guides');
 
-const loadTutorials = () =>
-  import(/* webpackChunkName: "discover-tutorials" */ './Tutorials');
+const loadTutorialList = () =>
+  import(
+    /* webpackChunkName: "discover-tutorials" */ './tutorials/TutorialList'
+  );
+const loadTutorialPage = () =>
+  import(
+    /* webpackChunkName: "tutorials-details-page" */ './tutorials/Tutorial'
+  );
 
 const loadWorkingGroups = () =>
   import(/* webpackChunkName: "discover-working-groups" */ './WorkingGroups');
@@ -16,24 +22,19 @@ const loadWorkingGroups = () =>
 const loadAbout = () =>
   import(/* webpackChunkName: "discover-about" */ './About');
 
-const loadTutorial = () =>
-  import(
-    /* webpackChunkName: "tutorials-details-page" */ './tutorials/Tutorial'
-  );
-
 const Guides = lazy(loadGuides);
-const Tutorials = lazy(loadTutorials);
+const TutorialList = lazy(loadTutorialList);
+const TutorialPage = lazy(loadTutorialPage);
 const WorkingGroups = lazy(loadWorkingGroups);
 const About = lazy(loadAbout);
-const TutorialDetailsPage = lazy(loadTutorial);
 
 const Discover: FC<Record<string, never>> = () => {
   useEffect(() => {
     loadGuides()
-      .then(loadTutorials)
+      .then(loadTutorialList)
       .then(loadWorkingGroups)
       .then(loadAbout)
-      .then(loadTutorials);
+      .then(loadTutorialPage);
   }, []);
 
   const { path } = useRouteMatch();
@@ -48,7 +49,7 @@ const Discover: FC<Record<string, never>> = () => {
         }
       >
         <Frame title={null}>
-          <TutorialDetailsPage />
+          <TutorialPage />
         </Frame>
       </Route>
       <DiscoverPage>
@@ -60,7 +61,7 @@ const Discover: FC<Record<string, never>> = () => {
           </Route>
           <Route exact path={path + discover({}).tutorials.template}>
             <Frame title="Tutorials">
-              <Tutorials />
+              <TutorialList />
             </Frame>
           </Route>
           <Route exact path={path + discover({}).workingGroups.template}>
